+++ lib/punchblock/event/dtmf.rb
@@ -3,25 +3,6 @@ module Punchblock
     class DTMF < Event
       register :dtmf, :core
 
-      ##
-      # Create a DTMF event
-      #
-      # @param [Hash] options
-      # @option options [String, Optional] :signal the DTMF signal received
-      #
-      # @return [Event::DTMF] a formatted Rayo DTMF event
-      #
-      def self.new(options = {})
-        super().tap do |new_node|
-          case options
-          when Nokogiri::XML::Node
-            new_node.inherit options
-          when Hash
-            options.each_pair { |k,v| new_node.send :"#{k}=", v }
-          end
-        end
-      end
-
       def signal
         read_attr :signal
       end